[REF] website: clean up website utils dead code
This commit removes duplicated utils and legacy functions that are no
longer used. It also relocates non-generic functions that are only used
in one place.
