+++ addons/website/static/src/builder/plugins/form/utils.js
@@ -12,33 +12,6 @@ export const VISIBILITY_DATASET = [
 ];
 
 /**
- * Returns the parsed data coming from the data-for element for the given form.
- * TODO Note that we should rely on the same util as the website form interaction.
- * Maybe this will need to be deleted.
- *
- * @param {string} formId
- * @param {HTMLElement} parentEl
- * @returns {Object|undefined} the parsed data
- */
-export function getParsedDataFor(formId, parentEl) {
-    const dataForEl = parentEl.querySelector(`[data-for='${formId}']`);
-    if (!dataForEl) {
-        return;
-    }
-    return JSON.parse(
-        dataForEl.dataset.values
-            // replaces `True` by `true` if they are after `,` or `:` or `[`
-            .replace(/([,:[]\s*)True/g, "$1true")
-            // replaces `False` and `None` by `""` if they are after `,` or `:` or `[`
-            .replace(/([,:[]\s*)(False|None)/g, '$1""')
-            // replaces the `'` by `"` if they are before `,` or `:` or `]` or `}`
-            .replace(/'(\s*[,:\]}])/g, '"$1')
-            // replaces the `'` by `"` if they are after `{` or `[` or `,` or `:`
-            .replace(/([{[:,]\s*)'/g, '$1"')
-    );
-}
-
-/**
  * Returns a field object
  *
  * @param {string} type the type of the field
