+++ addons/website/static/src/client_actions/website_preview/publish_website_systray_item.js
@@ -6,10 +6,44 @@ import { CheckBox } from "@web/core/checkbox/checkbox";
 import { useService, useBus } from "@web/core/utils/hooks";
 import { Component, useState } from "@odoo/owl";
 import { OptimizeSEODialog } from "@website/components/dialog/seo";
-import { checkAndNotifySEO } from "@website/js/utils";
 import { RelativePublishTime } from "./relative_publish_time";
 
 const websiteSystrayRegistry = registry.category("website_systray");
+
+/**
+ * Checks SEO data and notifies if either the page title or description is not
+ * set.
+ *
+ * @param {Object} seo_data - The SEO data to check.
+ * @param {Component} OptimizeSEODialog - Dialog to be displayed
+ * @param {Object} services - Services object which will be used to display
+ * notifications and dialog.
+ */
+export function checkAndNotifySEO(seo_data, OptimizeSEODialog, services) {
+    if (seo_data) {
+        let message;
+        if (!seo_data.website_meta_title) {
+            message = _t("Page title not set.");
+        } else if (!seo_data.website_meta_description) {
+            message = _t("Page description not set.");
+        }
+        if (message) {
+            const closeNotification = services.notification.add(message, {
+                type: "warning",
+                sticky: false,
+                buttons: [
+                    {
+                        name: _t("Optimize SEO"),
+                        onClick: () => {
+                            services.dialog.add(OptimizeSEODialog);
+                            closeNotification();
+                        },
+                    },
+                ],
+            });
+        }
+    }
+}
 
 export class PublishSystrayItem extends Component {
     static template = "website.PublishSystrayItem";
