+++ addons/website/static/src/js/utils.js
@@ -1,6 +1,4 @@
-import { intersection } from "@web/core/utils/arrays";
-import { _t, appTranslateFn } from "@web/core/l10n/translation";
-import { renderToElement } from "@web/core/utils/render";
+import { appTranslateFn } from "@web/core/l10n/translation";
 import { App, Component } from "@odoo/owl";
 import { getTemplate } from "@web/core/templates";
 import { UrlAutoComplete } from "@website/components/autocomplete_with_pages/url_autocomplete";
@@ -85,199 +83,6 @@ function autocompleteWithPages(input, options = {}, env = undefined) {
         owlApp.destroy();
         container.remove();
     };
-}
-
-/**
- * @param {jQuery} $element
- * @param {jQuery} [$excluded]
- */
-function onceAllImagesLoaded($element, $excluded) {
-    var defs = Array.from($element.find("img").addBack("img")).map((img) => {
-        if (img.complete || ($excluded && ($excluded.is(img) || $excluded.has(img).length))) {
-            return; // Already loaded
-        }
-        var def = new Promise(function (resolve, reject) {
-            $(img).one("load", function () {
-                resolve();
-            });
-        });
-        return def;
-    });
-    return Promise.all(defs);
-}
-
-/**
- * @deprecated
- * @todo create Dialog.prompt instead of this
- */
-function prompt(options, _qweb) {
-    /**
-     * A bootstrapped version of prompt() albeit asynchronous
-     * This was built to quickly prompt the user with a single field.
-     * For anything more complex, please use editor.Dialog class
-     *
-     * Usage Ex:
-     *
-     * website.prompt("What... is your quest?").then(function (answer) {
-     *     arthur.reply(answer || "To seek the Holy Grail.");
-     * });
-     *
-     * website.prompt({
-     *     select: "Please choose your destiny",
-     *     init: function () {
-     *         return [ [0, "Sub-Zero"], [1, "Robo-Ky"] ];
-     *     }
-     * }).then(function (answer) {
-     *     mame_station.loadCharacter(answer);
-     * });
-     *
-     * @param {Object|String} options A set of options used to configure the prompt or the text field name if string
-     * @param {String} [options.window_title=''] title of the prompt modal
-     * @param {String} [options.input] tell the modal to use an input text field, the given value will be the field title
-     * @param {String} [options.textarea] tell the modal to use a textarea field, the given value will be the field title
-     * @param {String} [options.select] tell the modal to use a select box, the given value will be the field title
-     * @param {Object} [options.default=''] default value of the field
-     * @param {Function} [options.init] optional function that takes the `field` (enhanced with a fillWith() method) and the `dialog` as parameters [can return a promise]
-     */
-    if (typeof options === "string") {
-        options = {
-            text: options,
-        };
-    }
-    if (typeof _qweb === "undefined") {
-        _qweb = "website.prompt";
-    }
-    options = Object.assign(
-        {
-            window_title: "",
-            field_name: "",
-            default: "", // dict notation for IE<9
-            init: function () {},
-            btn_primary_title: _t("Create"),
-            btn_secondary_title: _t("Cancel"),
-        },
-        options || {}
-    );
-
-    var type = intersection(Object.keys(options), ["input", "textarea", "select"]);
-    type = type.length ? type[0] : "input";
-    options.field_type = type;
-    options.field_name = options.field_name || options[type];
-
-    var def = new Promise(function (resolve, reject) {
-        var dialog = $(renderToElement(_qweb, options)).appendTo("body");
-        options.$dialog = dialog;
-        var field = dialog.find(options.field_type).first();
-        field.val(options["default"]); // dict notation for IE<9
-        field.fillWith = function (data) {
-            if (field.is("select")) {
-                var select = field[0];
-                data.forEach(function (item) {
-                    select.options[select.options.length] = new window.Option(item[1], item[0]);
-                });
-            } else {
-                field.val(data);
-            }
-        };
-        var init = options.init(field, dialog);
-        Promise.resolve(init).then(function (fill) {
-            if (fill) {
-                field.fillWith(fill);
-            }
-            dialog.modal("show");
-            field.focus();
-            dialog.on("click", ".btn-primary", function () {
-                var backdrop = $(".modal-backdrop");
-                resolve({ val: field.val(), field: field, dialog: dialog });
-                dialog.modal("hide").remove();
-                backdrop.remove();
-            });
-        });
-        dialog.on("hidden.bs.modal", function () {
-            var backdrop = $(".modal-backdrop");
-            reject();
-            dialog.remove();
-            backdrop.remove();
-        });
-        if (field.is('input[type="text"], select')) {
-            field.keypress(function (e) {
-                if (e.key === "Enter") {
-                    e.preventDefault();
-                    dialog.find(".btn-primary").trigger("click");
-                }
-            });
-        }
-    });
-
-    return def;
-}
-
-function websiteDomain(self) {
-    var websiteID;
-    self.trigger_up("context_get", {
-        callback: function (ctx) {
-            websiteID = ctx["website_id"];
-        },
-    });
-    return ["|", ["website_id", "=", false], ["website_id", "=", websiteID]];
-}
-
-/**
- * Checks if the 2 given URLs are the same, to prevent redirecting uselessly
- * from one to another.
- * It will consider naked URL and `www` URL as the same URL.
- * It will consider `https` URL `http` URL as the same URL.
- *
- * @param {string} url1
- * @param {string} url2
- * @returns {Boolean}
- */
-function isHTTPSorNakedDomainRedirection(url1, url2) {
-    try {
-        url1 = new URL(url1).host;
-        url2 = new URL(url2).host;
-    } catch {
-        // Incorrect URL, `false` URL..
-        return false;
-    }
-    return url1 === url2 || url1.replace(/^www\./, "") === url2.replace(/^www\./, "");
-}
-
-export function sendRequest(route, params) {
-    function _addInput(form, name, value) {
-        const param = document.createElement("input");
-        param.setAttribute("type", "hidden");
-        param.setAttribute("name", name);
-        param.setAttribute("value", value);
-        form.appendChild(param);
-    }
-
-    const form = document.createElement("form");
-    form.setAttribute("action", route);
-    form.setAttribute("method", params.method || "POST");
-    // This is an exception for the 404 page create page button, in backend we
-    // want to open the response in the top window not in the iframe.
-    if (params.forceTopWindow) {
-        form.setAttribute("target", "_top");
-    }
-
-    if (odoo.csrf_token) {
-        _addInput(form, "csrf_token", odoo.csrf_token);
-    }
-
-    for (const key in params) {
-        const value = params[key];
-        if (Array.isArray(value) && value.length) {
-            for (const val of value) {
-                _addInput(form, key, val);
-            }
-        } else {
-            _addInput(form, key, value);
-        }
-    }
-
-    document.body.appendChild(form);
-    form.submit();
 }
 
 /**
@@ -383,25 +188,6 @@ async function _exportToPNG(src, format) {
 }
 
 /**
- * Bootstraps an "empty" Google Maps iframe.
- *
- * @returns {HTMLIframeElement}
- */
-export function generateGMapIframe() {
-    const iframeEl = document.createElement("iframe");
-    iframeEl.classList.add("s_map_embedded", "o_not_editable");
-    iframeEl.setAttribute("width", "100%");
-    iframeEl.setAttribute("height", "100%");
-    iframeEl.setAttribute("frameborder", "0");
-    iframeEl.setAttribute("scrolling", "no");
-    iframeEl.setAttribute("marginheight", "0");
-    iframeEl.setAttribute("marginwidth", "0");
-    iframeEl.setAttribute("src", "about:blank");
-    iframeEl.setAttribute("aria-label", _t("Map"));
-    return iframeEl;
-}
-
-/**
  * Generates a Google Maps URL based on the given parameter.
  *
  * @param {DOMStringMap} dataset
@@ -420,30 +206,13 @@ export function generateGMapLink(dataset) {
 }
 
 /**
- * Checks if the edited content is currently previewed as in a mobile device.
- *
- * @param {Object} self - context object ("this")
- * @returns {boolean}
- */
-function isMobile(self) {
-    let isMobile;
-    self.trigger_up("service_context_get", {
-        callback: (ctx) => {
-            isMobile = ctx["isMobile"];
-        },
-    });
-
-    return isMobile;
-}
-
-/**
  * Returns the parsed data coming from the data-for element for the given form.
  *
  * @param {string} formId
  * @param {HTMLElement} parentEl
  * @returns {Object|undefined} the parsed data
  */
-function getParsedDataFor(formId, parentEl) {
+export function getParsedDataFor(formId, parentEl) {
     const dataForEl = parentEl.querySelector(`[data-for='${formId}']`);
     if (!dataForEl) {
         return;
@@ -482,41 +251,6 @@ export function cloneContentEls(content, keepScripts = false) {
         copyFragment.querySelectorAll("script").forEach((scriptEl) => scriptEl.remove());
     }
     return copyFragment;
-}
-
-/**
- * Checks SEO data and notifies if either the page title or description is not
- * set.
- *
- * @param {Object} seo_data - The SEO data to check.
- * @param {Component} OptimizeSEODialog - Dialog to be displayed
- * @param {Object} services - Services object which will be used to display
- * notifications and dialog.
- */
-export function checkAndNotifySEO(seo_data, OptimizeSEODialog, services) {
-    if (seo_data) {
-        let message;
-        if (!seo_data.website_meta_title) {
-            message = _t("Page title not set.");
-        } else if (!seo_data.website_meta_description) {
-            message = _t("Page description not set.");
-        }
-        if (message) {
-            const closeNotification = services.notification.add(message, {
-                type: "warning",
-                sticky: false,
-                buttons: [
-                    {
-                        name: _t("Optimize SEO"),
-                        onClick: () => {
-                            services.dialog.add(OptimizeSEODialog);
-                            closeNotification();
-                        },
-                    },
-                ],
-            });
-        }
-    }
 }
 
 /**
@@ -570,18 +304,8 @@ patch(urlUtils, {
 export default {
     loadAnchors: loadAnchors,
     autocompleteWithPages: autocompleteWithPages,
-    onceAllImagesLoaded: onceAllImagesLoaded,
-    prompt: prompt,
-    sendRequest: sendRequest,
-    websiteDomain: websiteDomain,
-    isHTTPSorNakedDomainRedirection: isHTTPSorNakedDomainRedirection,
     svgToPNG: svgToPNG,
     webpToPNG: webpToPNG,
-    generateGMapIframe: generateGMapIframe,
     generateGMapLink: generateGMapLink,
-    isMobile: isMobile,
-    getParsedDataFor: getParsedDataFor,
-    cloneContentEls: cloneContentEls,
-    checkAndNotifySEO: checkAndNotifySEO,
     slugify: slugify,
 };
