+++ addons/website/static/src/snippets/s_map/map.js
@@ -1,7 +1,26 @@
 import { Interaction } from "@web/public/interaction";
 import { registry } from "@web/core/registry";
+import { _t } from "@web/core/l10n/translation";
+import { generateGMapLink } from "@website/js/utils";
 
-import { generateGMapLink, generateGMapIframe } from "@website/js/utils";
+/**
+ * Bootstraps an "empty" Google Maps iframe.
+ *
+ * @returns {HTMLIframeElement}
+ */
+export function generateGMapIframe() {
+    const iframeEl = document.createElement("iframe");
+    iframeEl.classList.add("s_map_embedded", "o_not_editable");
+    iframeEl.setAttribute("width", "100%");
+    iframeEl.setAttribute("height", "100%");
+    iframeEl.setAttribute("frameborder", "0");
+    iframeEl.setAttribute("scrolling", "no");
+    iframeEl.setAttribute("marginheight", "0");
+    iframeEl.setAttribute("marginwidth", "0");
+    iframeEl.setAttribute("src", "about:blank");
+    iframeEl.setAttribute("aria-label", _t("Map"));
+    return iframeEl;
+}
 
 export class Map extends Interaction {
     static selector = ".s_map";
